Fix wysiwyg fields being added to a group
Because a wysiwyg field's id can't have a '/' and is sanitized, the
field name may be different than the id. We pass the id and name fields
via data separately and use them independently of each other.

Fixes issue where adding a new wysiwyg row to a group was not saving the
 value until after the page was reloaded and the value reentered.